wip: defined structure
