--- src/components/CustomModal/index.ts
@@ -1,1 +1,0 @@
-export * from "./CustomModal";
