+++ src/pages/games/chess/index.ts
