+++ src/pages/games/electronic/Electronic.tsx
