+++ src/pages/games/electronic/index.ts
