+++ src/pages/games/fishing/fishing.tsx
