+++ src/pages/games/fishing/index.ts
