+++ src/pages/games/lottery/Lottery.tsx
