+++ src/pages/games/lottery/index.ts
