+++ src/pages/games/video/index.ts
