+++ src/pages/games/video/video.tsx
